refactor: move boundaries to ports folder
