+++ src/sp-proxy/src/interface-adapters/delivery/AddTrFromMetadataController.ts
@@ -5,7 +5,7 @@ import { IRequest } from '@sp-proxy/interface-adapters/protocols/IRequest'
 import { IValidator } from '@sp-proxy/interface-adapters/protocols/IValidator'
 import { AddTrFromMetadataUseCaseParams } from '@sp-proxy/use-cases/io-models/AddTrFromMetadataUseCaseParams'
 import { IRequestModel } from '@sp-proxy/use-cases/io-models/IRequestModel'
-import { InputBoundary } from '@sp-proxy/use-cases/protocols/InputBoundary'
+import { InputBoundary } from '@sp-proxy/use-cases/ports/InputBoundary'
 
 export class AddTrFromMetadataController implements IController {
   constructor(
