+++ src/sp-proxy/src/interface-adapters/delivery/AddTrFromMetadataPresenter.ts
@@ -1,7 +1,7 @@
 import { IAddTrFromMetadataResponse } from '@sp-proxy/interface-adapters/protocols/IAddTrFromMetadataResponse'
 import { IDeliveryMapper } from '@sp-proxy/interface-adapters/protocols/IDeliveryMapper'
 import { IResponse } from '@sp-proxy/interface-adapters/protocols/IResponse'
-import { OutputBoundary } from '@sp-proxy/use-cases/protocols/OutputBoundary'
+import { OutputBoundary } from '@sp-proxy/use-cases/ports/OutputBoundary'
 import { IResponseModel } from '@sp-proxy/use-cases/io-models/IResponseModel'
 import { SuccessResponseUseCaseParams } from '@sp-proxy/use-cases/io-models/SuccessResponseUseCaseParams'
 import { EventEmitter } from 'stream'
