+++ src/sp-proxy/src/use-cases/ports/OutputBoundary.ts
